+++ Sound/Sound_test/test.cpp
@@ -1,9 +1,9 @@
 #include "pch.h"
 
-#include "SoundEngine.h"
+#include "AudioEngine.h"
 #include "RingBuffer.h"
 #include "TestPlugins.h"
-#include "SoundBank.h"
+#include "AudioBank.h"
 #include <thread>
 #include <cmath>
